reorganizing everything, there might be a couple o' bugs
