+++ __init__.py
@@ -1,0 +1,3 @@
+import os
+__all__ = [file.split('.')[0] for file in os.listdir('scripts')]
+
