+++ requirements.txt
@@ -1,0 +1,3 @@
+discord
+dotenv
+web3
